// pages/api/saveAutoBundleData.ts
export type GetByCollectionProducts = Array<{
  node: {
    products: {
      edges: Array<{
        node: {
          id: string;
          tags: Array<string>;
          priceRangeV2: {
            maxVariantPrice: {
              amount: string;
            };
          };
        };
      }>;
    };
  };
}>;